requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
python-dotenv==1.0.0